+++ src/main/java/com/revaure/mybankingapp/dao/AccountDAOImpl.java
@@ -16,9 +16,9 @@ public class AccountDAOImpl implements AccountDAO {
 	
 	private static final Logger log = Logger.getLogger(AccountDAOImpl.class.getName());
 
-	public String withdraw(Account account, long amount) {
+	public String withdraw(Account account, double amount) {
 		log.info("Hello this is an info message");
-		long newaccountbalance = account.getAccountbalance() - amount;
+		double newaccountbalance = account.getAccountbalance() - amount;
 		
 		if (newaccountbalance < 0) {
 			return "Cannot withdraw an amount larger than your account balance";
@@ -35,7 +35,7 @@ public class AccountDAOImpl implements AccountDAO {
 				stmt = connection.prepareStatement(sql);
 
 				// Add parameters for prepared statement
-				stmt.setLong(1, newaccountbalance);
+				stmt.setDouble(1, newaccountbalance);
 
 				stmt.setLong(2, account.getAccountnumber());
 
@@ -65,9 +65,9 @@ public class AccountDAOImpl implements AccountDAO {
 
 	}
 
-	public String transfer(long amount, Account fromaccount, long toaccountnumber) {
+	public String transfer(double amount, Account fromaccount, long toaccountnumber) {
 		
-		long newfromaccountbalance = fromaccount.getAccountbalance() - amount;
+		double newfromaccountbalance = fromaccount.getAccountbalance() - amount;
 
 		if (newfromaccountbalance < 0) {
 			return "Cannot withdraw an amount larger than your account balance";
@@ -84,7 +84,7 @@ public class AccountDAOImpl implements AccountDAO {
 				stmt = connection.prepareStatement(sql);
 
 				// Add parameters for prepared statement
-				stmt.setLong(1, newfromaccountbalance);
+				stmt.setDouble(1, newfromaccountbalance);
 
 				stmt.setLong(2, fromaccount.getAccountnumber());
 
@@ -114,7 +114,7 @@ public class AccountDAOImpl implements AccountDAO {
 				stmt = connection.prepareStatement(sql);
 
 				// Add parameters for prepared statement
-				stmt.setLong(1, amount);
+				stmt.setDouble(1, amount);
 
 				stmt.setLong(2, toaccountnumber);
 
@@ -143,8 +143,8 @@ public class AccountDAOImpl implements AccountDAO {
 		}
 	}
 
-	public String deposit(Account account, long amount) {
-		long newaccountbalance = account.getAccountbalance() + amount;
+	public String deposit(Account account, double amount) {
+		double newaccountbalance = account.getAccountbalance() + amount;
 
 		Connection connection = null;
 		PreparedStatement stmt = null;
@@ -158,7 +158,7 @@ public class AccountDAOImpl implements AccountDAO {
 			stmt = connection.prepareStatement(sql);
 
 			// Add parameters for prepared statement
-			stmt.setLong(1, newaccountbalance);
+			stmt.setDouble(1, newaccountbalance);
 
 			stmt.setLong(2, account.getAccountnumber());
 
@@ -245,7 +245,7 @@ public class AccountDAOImpl implements AccountDAO {
 					a.setOwnerid1(rs.getLong("ownerid1"));
 					a.setOwnerid2(rs.getLong("ownerid2"));
 					a.setRoutingnumber(rs.getLong("routingnumber"));
-					a.setAccountbalance(rs.getLong("accountbalance"));
+					a.setAccountbalance(rs.getDouble("accountbalance"));
 					a.setAccounttype(rs.getString("accounttype"));
 					
 					accounts.add(a);
@@ -269,4 +269,5 @@ public class AccountDAOImpl implements AccountDAO {
 			return accounts;
 	
 	}
+
 }
