+++ src/main/java/com/revaure/mybankingapp/dao/CustomerDAO.java
@@ -9,8 +9,8 @@ public interface CustomerDAO {
 	
 	public Customer login(String username, String password);
 	public ArrayList<Customer> getCustomers();
-	public String withdraw(Account account, long amount);
-	public String transfer(long amount, Account fromaccount, long toaccountnumber);
-	public String deposit(Account account, long amount);
+	public String withdraw(Account account, double amount);
+	public String transfer(double amount, Account fromaccount, long toaccountnumber);
+	public String deposit(Account account, double amount);
 	
 }
